fix: compilation for non macos env
