+++ src-tauri/src/services/engine/macos.rs
@@ -1,110 +1,10 @@
-use crate::entities::{Engine, EngineInfo, EngineStatus, InstallationMethod, InstallationProgress, ColimaConfig};
-use bollard::Docker;
-use std::env;
+use crate::entities::{
+    ColimaConfig, InstallationMethod, InstallationProgress,
+};
 use std::process::Command;
-use tokio::sync::mpsc;
 use tauri::{AppHandle, Emitter};
-use tracing::{debug, warn, instrument};
-
-
-#[instrument(skip_all, err)]
-async fn connect_to_docker_with_local_defaults() -> Result<Docker, String> {
-    debug!("Trying to connect to Docker via local defaults");
-
-    let docker = Docker::connect_with_local_defaults().map_err(|e| format!("Failed to connect to Docker: {}", e))?;
-
-    if let Ok(_) = docker.info().await {
-        debug!("Successfully connected to Docker via local defaults");
-        return Ok(docker);
-    }
-    warn!("local defaults connection failed, trying fallback");
-    Err("Failed to connect to Docker via local defaults".to_string())
-}
-
-#[instrument(skip_all, err)]
-async fn connect_to_docker_using_different_contexts() -> Result<Docker, String> {
-    debug!("Trying to connect to Docker via different contexts");
-
-    let context_output = Command::new("docker")
-        .arg("context")
-        .arg("ls")
-        .arg("--format")
-        .arg("{{.DockerEndpoint}}")
-        .output();
-
-    if let Ok(output) = context_output {
-        if output.status.success() {
-            let output_str = String::from_utf8_lossy(&output.stdout).trim().to_string();
-            let socket_paths = output_str.lines();
-
-            for socket_path in socket_paths {
-                if socket_path.starts_with("unix://") {
-                    let socket_path = socket_path.trim_start_matches("unix://");
-                    debug!("Current Docker context socket: {}", socket_path);
-
-                    // Set DOCKER_HOST environment variable for this process
-                    env::set_var("DOCKER_HOST", format!("unix://{}", socket_path));
-
-                    if let Ok(docker) = Docker::connect_with_local_defaults() {
-                        if let Ok(_) = docker.info().await {
-                            debug!("Successfully connected to Docker via context socket");
-                            return Ok(docker);
-                        }
-                    }
-                } else {
-                    warn!("Current Docker context socket is not a Unix socket: {}", socket_path);
-                }
-
-            }
-        }
-    }
-
-    Err("Failed to connect to Docker via context socket".to_string())
-}
-
-#[instrument(skip_all, err)]
-pub async fn connect_to_docker_macos() -> Result<Docker, String> {
-    // First, check if DOCKER_HOST environment variable is set
-    if let Ok(docker) = connect_to_docker_with_local_defaults().await {
-        return Ok(docker);
-    }
-    debug!("local defaults connection failed, trying fallback");
-
-
-    // Try to get the current Docker context
-    if let Ok(docker) = connect_to_docker_using_different_contexts().await {
-        return Ok(docker);
-    }
-    debug!("context connection failed, trying fallback");
-
-    // Last resort: try the default connection method
-    debug!("Trying default Docker connection method");
-    Docker::connect_with_local_defaults().map_err(|e| format!("Failed to connect to Docker: {}", e))
-}
-
-#[instrument(skip_all, err)]
-pub async fn is_docker_command_available() -> Result<bool, String> {
-    debug!("Checking if Docker command is available");
-    fn is_docker_installed(output_str: &str) -> bool {
-        output_str.to_lowercase().contains("version")
-    }
-
-    let output = Command::new("docker")
-        .arg("--version")
-        .output()
-        .map_err(|e| format!("Failed to execute docker command: {}", e))?;
-
-    if !output.status.success() {
-        debug!("Docker command is not available, output: {}", String::from_utf8_lossy(&output.stderr));
-        return Ok(false);
-    }
-
-    let output_str = String::from_utf8_lossy(&output.stdout);
-    debug!("Docker command is available, output: {}", output_str);
-    let result = is_docker_installed(&output_str);
-    debug!("Docker command is available, result: {}", result);
-    Ok(result)
-}
+use tokio::sync::mpsc;
+use tracing::{debug, instrument};
 
 #[instrument(skip_all, err)]
 pub async fn is_homebrew_available() -> Result<bool, String> {
@@ -153,35 +53,11 @@ pub async fn is_colima_available() -> Result<bool, String> {
         return Ok(false);
     }
 
-    let version = String::from_utf8_lossy(&version_output.stdout).trim().to_string();
+    let version = String::from_utf8_lossy(&version_output.stdout)
+        .trim()
+        .to_string();
     debug!("Colima is available, version: {}", version);
     Ok(true)
-}
-
-#[instrument(skip_all, err)]
-pub async fn create_engine() -> Result<Engine, String> {
-    debug!("Creating an engine instance");
-
-    if !is_docker_command_available().await? {
-        debug!("Docker command is not available, creating an engine instance with unknown status");
-        return Ok(Engine {
-            engine_status: EngineStatus::Unknown,
-            docker: None,
-        });
-    }
-
-    if let Ok(docker) = connect_to_docker_macos().await {
-        debug!("Docker command is available, creating an engine instance with running status");
-        return Ok(Engine {
-            engine_status: EngineStatus::Running(EngineInfo::Docker),
-            docker: Some(docker),
-        });
-    }
-
-    Ok(Engine {
-        engine_status: EngineStatus::Unknown,
-        docker: None,
-    })
 }
 
 #[instrument(skip_all, err)]
@@ -192,12 +68,8 @@ pub async fn install_colima(
     debug!("Starting Colima installation via {:?}", method);
 
     match method {
-        InstallationMethod::Homebrew => {
-            install_colima_via_homebrew(app_handle).await
-        }
-        InstallationMethod::Binary => {
-            Err("Binary installation not yet implemented".to_string())
-        }
+        InstallationMethod::Homebrew => install_colima_via_homebrew(app_handle).await,
+        InstallationMethod::Binary => Err("Binary installation not yet implemented".to_string()),
     }
 }
 
@@ -227,7 +99,9 @@ async fn install_colima_via_homebrew(app_handle: AppHandle) -> Result<(), String
         progress.step = "Installing Colima...".to_string();
         progress.message = "Installing Colima via Homebrew".to_string();
         progress.percentage = 30;
-        progress.logs.push("[INFO] Installing Colima via Homebrew".to_string());
+        progress
+            .logs
+            .push("[INFO] Installing Colima via Homebrew".to_string());
         let _ = tx_clone.send(progress.clone()).await;
 
         let install_result = install_colima_packages().await;
@@ -240,7 +114,9 @@ async fn install_colima_via_homebrew(app_handle: AppHandle) -> Result<(), String
             return Err(e.clone());
         }
 
-        progress.logs.push("[INFO] Colima installation completed successfully".to_string());
+        progress
+            .logs
+            .push("[INFO] Colima installation completed successfully".to_string());
         let _ = tx_clone.send(progress.clone()).await;
 
         Ok(())
@@ -255,7 +131,8 @@ async fn install_colima_via_homebrew(app_handle: AppHandle) -> Result<(), String
     });
 
     // Wait for installation to complete
-    let result = installation_handle.await
+    let result = installation_handle
+        .await
         .map_err(|e| format!("Installation task failed: {}", e))?
         .and_then(|r| Ok(r));
 
@@ -299,10 +176,7 @@ async fn install_colima_packages() -> Result<(), String> {
 }
 
 #[instrument(skip_all, err)]
-pub async fn start_colima_vm(
-    app_handle: AppHandle,
-    config: ColimaConfig,
-) -> Result<(), String> {
+pub async fn start_colima_vm(app_handle: AppHandle, config: ColimaConfig) -> Result<(), String> {
     debug!("Starting Colima VM with config: {:?}", config);
 
     // Create a channel for progress updates
@@ -327,7 +201,9 @@ pub async fn start_colima_vm(
         progress.step = "Checking Colima status...".to_string();
         progress.message = "Verifying current VM status".to_string();
         progress.percentage = 20;
-        progress.logs.push("[INFO] Checking Colima status".to_string());
+        progress
+            .logs
+            .push("[INFO] Checking Colima status".to_string());
         let _ = tx_clone.send(progress.clone()).await;
 
         let status_result = check_colima_status().await;
@@ -335,7 +211,9 @@ pub async fn start_colima_vm(
             progress.step = "Colima already running".to_string();
             progress.message = "Colima VM is already running".to_string();
             progress.percentage = 100;
-            progress.logs.push("[INFO] Colima VM is already running".to_string());
+            progress
+                .logs
+                .push("[INFO] Colima VM is already running".to_string());
             let _ = tx_clone.send(progress.clone()).await;
             return Ok(());
         }
@@ -361,7 +239,9 @@ pub async fn start_colima_vm(
         progress.step = "Validating startup...".to_string();
         progress.message = "Testing Docker connectivity".to_string();
         progress.percentage = 80;
-        progress.logs.push("[INFO] Validating Colima startup".to_string());
+        progress
+            .logs
+            .push("[INFO] Validating Colima startup".to_string());
         let _ = tx_clone.send(progress.clone()).await;
 
         // Wait a bit for the VM to fully start
@@ -380,7 +260,9 @@ pub async fn start_colima_vm(
         progress.step = "VM startup complete".to_string();
         progress.message = "Colima VM is running successfully".to_string();
         progress.percentage = 100;
-        progress.logs.push("[INFO] Colima VM started successfully".to_string());
+        progress
+            .logs
+            .push("[INFO] Colima VM started successfully".to_string());
         let _ = tx_clone.send(progress.clone()).await;
 
         Ok(())
@@ -395,7 +277,8 @@ pub async fn start_colima_vm(
     });
 
     // Wait for VM startup to complete
-    let result = vm_handle.await
+    let result = vm_handle
+        .await
         .map_err(|e| format!("VM startup task failed: {}", e))?
         .and_then(|r| Ok(r));
