+++ src-tauri/src/services/mod.rs
@@ -1,10 +1,10 @@
 mod config;
 mod containers;
+pub(crate) mod engine;
 mod images;
 mod networks;
 mod updater;
 mod volumes;
-pub(crate) mod engine;
 
 pub use config::*;
 pub use containers::*;
